working on shopping and stuff

+++ lab2/ProductClass.cs
@@ -11,7 +11,7 @@ namespace lab2
     public class Product
     {
         public string Name { get; set; }
-        public decimal PricePerUnit { get; set; }
+        public int PricePerUnit { get; set; }
         public int Quantity { get; set; }
 
         public override string ToString()
